+++ pyrenew/latent/infectionswithfeedback.py
@@ -169,7 +169,7 @@ class InfectionsWithFeedback(RandomVariable):
         )
 
         if inf_feedback_strength.ndim == Rt.ndim - 1:
-            inf_feedback_strength = jnp.expand_dims(inf_feedback_strength, 0)
+            inf_feedback_strength = inf_feedback_strength[jnp.newaxis]
 
         # Making sure inf_feedback_strength spans the Rt length
         if inf_feedback_strength.shape[0] == 1:
